Milo anda apenas onde pode e da forma adequada

+++ lab07/Dungeons&Wumpus/src/mc322/game/entitiesCharacters/Heros.java
@@ -2,7 +2,7 @@ package mc322.game.entitiesCharacters;
 
 public abstract class Heros extends Character{
 
-	public Heros(int i, int j,int elevation) {
+	public Heros(int i, int j,double elevation) {
 		super(i, j, elevation);
 	}
